refactor: extract pure canvas spec/evalExpr for renderer use
